part3/requirement.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
cryptography==37.0.4
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1
django-allauth==0.51.0
django-extensions==3.2.0
django-rest-auth==0.9.5
django-restframework==0.0.1
djangorestframework==3.13.1
executing==0.10.0
idna==3.3
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.6
oauthlib==3.2.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.4.0
traitlets==5.3.0
tzdata==2022.2
urllib3==1.26.12
wcwidth==0.2.5
